requirements.txt
numpy
pandas
Pillow
requests
tqdm
wget
medclip
opencv-python>=4.4.0.42
nltk>=3.7
scikit_learn>=1.1.2
textaugment>=1.3.4
timm>=0.6.11
torch>=1.12.1
torchvision>=0.13.1
transformers>=4.23.1,<=4.24.0
gradio==3.35.2
